refactor: extract type analysis and resolver utilities
Extract type analysis logic into separate modules for better code organization:
- Move covariant subtype checking to type-analysis.ts
- Add type resolver utilities to type-reference.ts
- Simplify validator.ts by removing duplicated type resolution logic
